requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
fastapi==0.115.5
flet==0.26.0
flet-cli==0.26.0
flet-core==0.24.1
flet-desktop==0.26.0
flet-runtime==0.24.1
flet-web==0.26.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.2.0
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==23.2
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic_core==2.27.0
pygame==2.6.1
Pygments==2.18.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==13.9.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==14.1
